Clean up imports by moving dataset class into dataset.py
